+++ app/airdrops/[slug]/page.tsx
@@ -1,17 +1,15 @@
-
 // import Categories from "@/components/airdrops/Categories";
 import Airdrops from "@/components/airdrops/Airdrops";
 import Categories from "@/components/airdrops/Categories";
 import CategoryMenu from "@/components/airdrops/CategoryMenu";
 
-
-import CardSection from "@/components/ui/cardSection";
+import CardSection from "@/components/home/cardSection";
 
 const page = () => {
   return (
     <div className="wrapper  pt-3">
       {/* <Categories /> */}
-     <CategoryMenu></CategoryMenu>
+      <CategoryMenu></CategoryMenu>
       {/* <Airdrops /> */}
       <CardSection></CardSection>
     </div>
