+++ app/page.tsx
@@ -2,16 +2,14 @@ import BannerAds from "@/components/home/BannerAds";
 import HotDrops from "@/components/home/HotDrops";
 import JoinUs from "@/components/home/JoinUs";
 
-import CardSection from "@/components/ui/cardSection";
-
+import CardSection from "@/components/home/cardSection";
 
 const page = () => {
   return (
     <div className="wrapper  pt-3">
       <BannerAds />
       <HotDrops />
-      {/* <RecentDrops /> */}
-     <CardSection></CardSection>
+      <CardSection></CardSection>
       <JoinUs />
     </div>
   );
